+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added
 - `CodePoint.toUnicodeNotation()` returns the standard Unicode notation of a code point, e.g. `U+1F4E7`. 
+- `CharSequence.codePointCount()` variant without parameters.
 
 ## [0.8.0] - 2024-06-09
 ### Changed